Use `Locale::getDefault()` instead of `setlocale(..., 0)`
When using `intl` formatters, passing `C`[^C] as the locale is not valid for
ICU. This was acceptable until a breaking change in
[PHP](https://github.com/php/php-src/issues/12282), which remains broken in
PHP 8.1.x. Now, `Locale::getDefault()` is used to ensure valid locales are
passed.

[^C]: If the selected locale in Icinga Web is not installed on the system,
`setlocale()` defaults to `C`. A workaround without this patch is to install
the locale.